typechecker: Return type wrapped in an Optional when typechecking None
If it isn't already an Optional nor a WeakPtr. This means that we
will always try to infer the Type to an Optional<Type> instead
of a WeakPtr<Type>, but this seems like a good default.
If someone wants to return a WeakPtr<Type>, direct type hint is
required (previously we would just error out).